Partially fixes a problem with how parameters are handled for slices of compound models created from model instances, brought up in this comment:
https://github.com/astropy/astropy/pull/3231#issuecomment-67545043

This is still a bit of a hack, but not such a bad one, and far less
brittle than the temporary fix that was introduced in 2112c12 (I didn't
necessarily intend it to be temporary but I at least deeply hoped that a
better way could be found later--for now this is it).

This makes it possible to pass in additional class members that should
be added to classes generated from _CompoundModelMeta._from_operator,
and includes a helper function _model_operator that makes it easy to
make replacements for the standard operator functions (operator.add,
etc.) that can include additional members in the returned classes.  This
way, when taking slices, we can include the critical _slice_offset
attribute in the class creation as soon as it is needed.

Clearly this is only an internal detail right now, but I can envision
this being potentially useful to open this up, in some form, to a
documented feature for customizing the behavior of model expressions
(likely with the help of some context manager in which the expressions
are evaluated...)